Bug fixing and improving the unit test cases.
Found a couple things I missed in resetting the state post-load.
The only known issue left to resolve is that I think the animation_player_state
is going to need extra work to save and load the frame of the animation that is
playing when it is saved. I'm not sure if that's possible or not, but will
investigate.